requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.64
botocore==1.34.64
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
jmespath==1.0.1
multidict==6.0.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
yarl==1.9.4
